requirement.txt
streamlit
pypdf2
langchain
python-dotenv
chromadb
huggingface_hub
sentence-transformers